+++ src/VaultSharp/V1/AuthMethods/JWT/IJWTAuthMethod.cs
@@ -1,0 +1,9 @@
+﻿namespace VaultSharp.V1.AuthMethods.JWT
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public interface IJWTAuthMethod
+    {
+    }
+}