+++ CHANGELOG.md
@@ -4,6 +4,7 @@ FEATURES:
 
   * Add support for Azure Auth method login.
   * Add support for GoogleCloud Auth method login.
+  * Add support for JWT/OIDC Auth method login.
  
 BREAKING CHANGES: